Some restructuring, also added CourseEventBroadcastHelper
