+++ app/src/androidTest/java/aaron/watson/notekeeper/NoteCreationTest.java
@@ -7,9 +7,9 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import aaron.watson.notekeeper.course.CourseInfo;
-import aaron.watson.notekeeper.data.DatabaseManager;
+import aaron.watson.notekeeper.database.DatabaseManager;
 import aaron.watson.notekeeper.note.NoteInfo;
-import aaron.watson.notekeeper.note.NoteListActivity;
+import aaron.watson.notekeeper.android.activities.NoteListActivity;
 import static org.junit.Assert.*;
 import static android.support.test.espresso.Espresso.*;
 import static android.support.test.espresso.matcher.ViewMatchers.*;
