+++ app/src/main/java/aaron/watson/notekeeper/database/NoteKeeperDatabaseContract.java
@@ -1,4 +1,4 @@
-package aaron.watson.notekeeper.note;
+package aaron.watson.notekeeper.database;
 
 import android.provider.BaseColumns;
