+++ app/src/main/res/layout/content_note_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".note.NoteListActivity"
+    tools:context=".android.activities.NoteListActivity"
     tools:showIn="@layout/activity_note_list">
 
     <android.support.v7.widget.RecyclerView
