+++ app/src/main/res/menu/menu_note.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="aaron.watson.notekeeper.note.NoteActivity">
+    tools:context=".android.activities.NoteActivity">
     <item
         android:id="@+id/action_cancel"
         android:icon="@android:drawable/ic_menu_revert"
