+++ app/src/test/java/aaron/watson/notekeeper/DatabaseManagerTest.java
@@ -4,10 +4,8 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-import java.util.List;
-
 import aaron.watson.notekeeper.course.CourseInfo;
-import aaron.watson.notekeeper.data.DatabaseManager;
+import aaron.watson.notekeeper.database.DatabaseManager;
 import aaron.watson.notekeeper.note.NoteInfo;
 
 import static org.junit.Assert.*;
